% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM4TIF.R
\name{addTM2tif}
\alias{addTM2tif}
\title{Title}
\usage{
addTM2tif(tif_path, csv_path)
}
\arguments{
\item{tif_path}{path of tif file}

\item{csv_path}{path of csv file containing TM}
}
\value{
a raster with time attr
}
\description{
Title
}
\examples{
# addTM2tif('inst/data/raw/PML_V2_Ec.tif',
#           'inst/data/raw/PML_V2_Ec.csv')
}
