% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_array_layer.R
\name{bind_array_layer}
\alias{bind_array_layer}
\title{Merge layers of the same shape array}
\usage{
bind_array_layer(arr1, arr2)
}
\arguments{
\item{arr1}{A array}

\item{arr2}{A array}
}
\description{
Merge layers of the same shape array
}
\examples{
bind_array_layer(array(1:2, c(1, 2)), array(1:2, c(1, 2)))
}
