% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_delta}
\alias{cal_delta}
\title{Calculate the slope of the saturation vapor pressure curve at certain air
temperature Tair}
\usage{
cal_delta(Tair)
}
\arguments{
\item{Tair}{air temperature [degC]}
}
\value{
slope (delta) [kPa degC-1]
}
\description{
Calculate the slope of the saturation vapor pressure curve at certain air
temperature Tair
}
\examples{
cal_delta(20)
}
