% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_ea}
\alias{cal_ea}
\title{Calculate actual vapor pressure}
\usage{
cal_ea(Ta, RH = NULL, Tdew = NULL, Pa = NULL, q = NULL)
}
\arguments{
\item{Ta}{air temperature [degC]   (method_1 param)}

\item{RH}{relative humidity [%]    (method_1 param)}

\item{Tdew}{dew-point temperature  (method_2 param)}

\item{Pa}{pressure [kPa]           (method_2 param)}

\item{q}{specif humidity [kg kg-1] (method_3 param)}
}
\value{
actual vapor pressure [kPa]
}
\description{
Calculate actual vapor pressure
}
\examples{
cal_ea(q = 0.5)
}
