% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_Ep}
\alias{cal_Ep}
\title{Calculate potential evapotranspiration by Penman 1948 model}
\usage{
cal_Ep(Ta, Rn, U2, ea, G = NULL, Pa = NULL, gma = NULL)
}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Rn}{net radiation [W m-2]}

\item{U2}{2 meters wind speed [m s-1]}

\item{ea}{actual vapor pressure [kPa]}

\item{G}{soil heat flux [W m-2]}

\item{Pa}{pressure [kPa]}

\item{gma}{psychrometric constant [kPa degC-1]}
}
\value{
potential evapotranspiration calculated by Penman 1948 model [mm d-1]
}
\description{
Calculate potential evapotranspiration by Penman 1948 model
}
\examples{
cal_Ep(Ta = 20, Rn = 50, U2 = 5, ea = 0.3)
}
