% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_Ep_max}
\alias{cal_Ep_max}
\title{Calculate maximum value of potential evapotranspiration in dry environment}
\usage{
cal_Ep_max(Tdry, Rn, U2, G = NULL, Pa = NULL, gma = NULL)
}
\arguments{
\item{Tdry}{dry environment temperature [degC]}

\item{Rn}{net radiation [W m-2]}

\item{U2}{2 meters wind speed [m s-1]}

\item{G}{soil heat flux [W m-2]}

\item{Pa}{pressure [kPa]}

\item{gma}{psychrometric constant [kPa degC-1]}
}
\value{
max value of potential evapotranspiration in dry environment [mm d-1]
}
\description{
Calculate maximum value of potential evapotranspiration in dry environment
}
\examples{
cal_Ep_max(Tdry = 25, Rn = 50, U2 = 3)
}
