% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_es}
\alias{cal_es}
\title{Calculate saturation vapor pressure}
\usage{
cal_es(Tair)
}
\arguments{
\item{Tair}{air temperature [degC]}
}
\value{
saturation vapor pressure [kPa]
}
\description{
Calculate saturation vapor pressure
}
\examples{
cal_es(20)
}
