% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_ET_by_CR_model}
\alias{cal_ET_by_CR_model}
\title{Calculate evapotranspiration by calibration-free complementary relationship model}
\usage{
cal_ET_by_CR_model(
  Ta,
  Td,
  U2,
  Rn,
  Ep = NULL,
  G = NULL,
  gma = NULL,
  Pa = NULL,
  q = NULL,
  RH = NULL
)
}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Td}{dew-point temperature [degC]}

\item{U2}{2 meters wind speed [m s-1]}

\item{Rn}{net radiation [W m-2]}

\item{G}{soil heat flux [W m-2]}

\item{gma}{psychrometric constant [kPa degC-1]}

\item{Pa}{pressure [kPa]}

\item{q}{specific humidity [kg kg-1]}

\item{RH}{relative humidity [\%]}
}
\value{
evapotranspiration calculated by calibration-free CR model [mm d-1]
}
\description{
Calculate evapotranspiration by calibration-free complementary relationship model
}
\examples{
cal_ET_by_CR_model(Ta = 20, Td = 8, U2 = 3, Rn = 50)
}
