% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_Ew}
\alias{cal_Ew}
\title{Calculate evapotranspiration by Priestley-Taylor (PT) model}
\usage{
cal_Ew(Tw, Rn, G = NULL, Pa = NULL, gma = NULL)
}
\arguments{
\item{Tw}{wet environment temperature [degC]}

\item{Rn}{net radiation [W m-2]}

\item{G}{soil heat flux [W m-2]}

\item{Pa}{pressure [kPa]}

\item{gma}{psychrometric constant [kPa degC-1]}
}
\value{
evapotranspiration calculated by Priestley-Taylor model [mm d-1]
}
\description{
Calculate evapotranspiration by Priestley-Taylor (PT) model
}
\examples{
cal_Ew(Tw = 10, Rn = 50)
}
