% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_lambda}
\alias{cal_lambda}
\title{Calculate lantent heat of vaporization}
\usage{
cal_lambda(Tair)
}
\arguments{
\item{Tair}{air temperature [degC]}
}
\value{
lambda, latent heat of vaporization, about [2.45 MJ kg-1]
}
\description{
Calculate lantent heat of vaporization
}
\examples{
cal_lambda(20)
}
