% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_Tdry}
\alias{cal_Tdry}
\title{Calculate dry environment temperature}
\usage{
cal_Tdry(Twb, Pa = NULL, gma = NULL)
}
\arguments{
\item{Twb}{wet-bulb temperature [degC]}

\item{Pa}{pressure [kPa]}

\item{gma}{psychrometric constant [kPa degC-1]}
}
\value{
dry environment temperature [degC]
}
\description{
Calculate dry environment temperature
}
\examples{
cal_Tdry(15)
}
