% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cal_Twb}
\alias{cal_Twb}
\title{Calculate wet-bulb temperature}
\usage{
cal_Twb(VPD, Ta, Pa = 101.325)
}
\arguments{
\item{VPD}{vapor pressure deficit [kPa]}

\item{Ta}{air temperature [degC]}

\item{Pa}{air pressure [kPa]}
}
\value{
wet-bulb temperature [degC]
}
\description{
Calculate wet-bulb temperature
}
\examples{
cal_Twb(0.5, 20)
}
