% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{cal_Tws}
\alias{cal_Tws}
\title{Calculate wet surface temperature}
\usage{
cal_Tws(Ta, Rn, U2, VPD, Pa = 101.325, G = NULL)
}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Rn}{surface net radiation [W m-2]}

\item{U2}{2 meters wind speed [m s-1]}

\item{VPD}{vapor pressure deficit [kPa]}

\item{Pa}{air pressure [kPa]}

\item{G}{soil heat flux [W m-2]}
}
\value{
wet surface temperature [degC]
}
\description{
Calculate wet surface temperature
}
\examples{
cal_Tws(20, 50, 3, 0.5)
}
