% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_U10_to_U2}
\alias{cal_U10_to_U2}
\title{Converted 10m wind speed to 2m wind speed}
\usage{
cal_U10_to_U2(U10)
}
\arguments{
\item{U10}{10 meters wind speed [m s-1]}
}
\value{
2 meters wind speed [m s-1]
}
\description{
Converted 10m wind speed to 2m wind speed
}
\examples{
cal_U10_to_U2(5)
}
