% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_functions.R
\name{cal_U2}
\alias{cal_U2}
\title{Converted h(m) wind speed to 2(m) wind speed}
\arguments{
\item{Uz}{z meters wind speed [m s-1]}

\item{z}{measurement altitude of wind speed [m]}
}
\value{
2 meters wind speed [m s-1]
}
\description{
Converted h(m) wind speed to 2(m) wind speed
}
\examples{
cal_U2(6)
}
