% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{cal_VPD}
\alias{cal_VPD}
\title{Calculating vapor pressure deficit}
\usage{
cal_VPD(Ta, Pa = 101.325, ea = NULL, Tdew = NULL, q = NULL)
}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Pa}{air pressure [kPa]}

\item{ea}{actual vapour pressure [kPa]}

\item{Tdew}{dew-point temperature [degC]}

\item{q}{specific humidity [g g-1]}
}
\value{
vapor pressure deficit [kPa]
}
\description{
Calculating vapor pressure deficit
}
\examples{
cal_VPD(Ta = 20, ea = 1)
}
