% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{calib_alpha}
\alias{calib_alpha}
\title{Calibrating PT-alpha coefficients for widely wet surfaces}
\usage{
calib_alpha(Ta, Tws, VPD, Pa = 101.325)
}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Tws}{wet surface temperature [degC]}

\item{VPD}{vapor pressure deficit [kPa]}

\item{Pa}{air pressure [kPa]}
}
\value{
PT-alpha coefficients
}
\description{
Calibrating PT-alpha coefficients for widely wet surfaces
}
\examples{
calib_alpha(20, 10, 0.5)
}
