% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{ecm}
\alias{ecm}
\title{Elasticity Coefficient Method (ECM)}
\usage{
ecm(df, y_name, x_name, method = "lu2025")
}
\arguments{
\item{df}{a data.frame with input data}

\item{y_name}{y column name}

\item{x_name}{x column name(s)}

\item{method}{method used to calculate delta value}
}
\value{
a data.frame
}
\description{
Elasticity Coefficient Method (ECM)
}
\references{
Zhang et al., 2023, Nature Water
}
