% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_PT1972.R
\name{ET_PT1972}
\alias{ET_PT1972}
\title{Title Calculate evapotranspiration by Priestley-Taylor (PT) model}
\usage{
ET_PT1972(Tw, Rn, Pa = 101.325, alpha = 1.26, Ta = NULL, G = NULL)
}
\arguments{
\item{Tw}{wet environment temperature [degC]}

\item{Rn}{surface net radiation [W m-2]}

\item{Pa}{air pressure [kPa]}

\item{alpha}{Priestley-Taylor coefficient}

\item{Ta}{air temperature [degC]}

\item{G}{soil heat flux [W m-2]}
}
\value{
evapotranspiration calculated by Priestley-Taylor model [mm d-1]
}
\description{
Title Calculate evapotranspiration by Priestley-Taylor (PT) model
}
\examples{
ET_PT1972(10, 50)
}
