% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_separate.R
\name{exact_separate}
\alias{exact_separate}
\title{Enhanced separate function}
\usage{
exact_separate(
  d,
  .col_name,
  .col_value,
  .col_name_sep = "[^[:alnum:]]+",
  .col_value_sep = "[^[:alnum:]]+",
  .remove = TRUE,
  .convert = FALSE,
  .extra = "warn",
  .fill = "warn",
  ...
)
}
\arguments{
\item{d}{A data frame}

\item{.col_name}{column name of the col_name to separate}

\item{.col_value}{column name of the col_value to separate}

\item{.col_name_sep}{Separation of column names}

\item{.col_value_sep}{Separation of column values}

\item{.remove}{Same as the parameters of `separate`}

\item{.convert}{Same as the parameters of `separate`}

\item{.extra}{Same as the parameters of `separate`}

\item{.fill}{Same as the parameters of `separate`}

\item{...}{Same as the parameters of `separate`}
}
\description{
Enhanced separate function
}
\examples{
d <- data.frame(FORMAT = c('B:C', 'A:B:C', 'B:C:D', 'A:C', 'A:B'),
value = c('2:3', '1:2:3', '2:3:4', '1:3', '1:2'),tes = 1)
res <- exact_separate(d, .col_name = FORMAT, .col_value = value)
}
