% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlerTools.R
\name{get_NCBI_info}
\alias{get_NCBI_info}
\title{get NCBI sample information table}
\usage{
get_NCBI_info(sample_ID, tidy = T, project = F)
}
\arguments{
\item{sample_ID}{sample ID}

\item{tidy}{whether tidy the fieldvals table}

\item{project}{whether sample_ID is project_ID, `tidy` default to TRUE}
}
\value{
NCBI_info (list or data.frame)
}
\description{
get NCBI sample information table
}
\examples{
get_NCBI_info(sample_ID = 'ERX2746395')
}
