% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlerTools.R
\name{get_plantintron_data}
\alias{get_plantintron_data}
\title{Crawling Plant Intron Splicing Efficiency Database}
\usage{
get_plantintron_data(ath = "AT1G01830")
}
\arguments{
\item{ath}{Gene}
}
\value{
data.table
}
\description{
Crawling Plant Intron Splicing Efficiency Database
}
\examples{
get_plantintron_data()
}
