% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlerTools.R
\name{get_protein_homologs}
\alias{get_protein_homologs}
\title{get protein homologs from JGI Phytozome}
\usage{
get_protein_homologs(transcript)
}
\arguments{
\item{transcript}{query transcript name}
}
\value{
a data.table
}
\description{
get protein homologs from JGI Phytozome
}
\examples{
get_protein_homologs('Cre01.g005534.t1.2')
}
