% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlerTools.R
\name{GetZQData}
\alias{GetZQData}
\title{Get water level (Z) and runoff (Q) data of hydro station}
\usage{
GetZQData(stcd, sdate, edate = NULL)
}
\arguments{
\item{stcd}{hydrologic station}

\item{sdate}{start date e.m. 1970-01-01}

\item{edate}{end date}
}
\value{
a data.table with hydro data
}
\description{
Get water level (Z) and runoff (Q) data of hydro station
}
\examples{
GetZQData(60106000, sdate = '1970-01-01', edate = '2022-08-30')
}
