% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRunoff.R
\name{ggRunoff}
\alias{ggRunoff}
\title{Drawing rainfall runoff map based on ggplot2}
\usage{
ggRunoff(
  d,
  date,
  Q,
  prcp,
  times = 1,
  facet = F,
  col_q = NULL,
  ymax = NULL,
  prcp_color = NULL,
  axis1_lim = NULL,
  Axis1 = "Axis1",
  Axis2 = "Axis2",
  line_size = 1
)
}
\arguments{
\item{d}{input data [data.frame]}

\item{date}{Column name of date [character]}

\item{Q}{column name of runoff [character]}

\item{prcp}{column name of precipitation [character]}

\item{times}{magnification precipitation}

\item{facet}{if TRUE, facet_wrap is used}

\item{col_q}{Column name of `color type` [character]}

\item{ymax}{Y-axis maximum, also second abscissa axis position}

\item{prcp_color}{fill color for rainfall}

\item{axis1_lim}{limitations of axis1}

\item{Axis1}{label of axis1}

\item{Axis2}{label of axis2}

\item{line_size}{size of geom_line (runoff)}
}
\value{
rainfall runoff map based on ggplot2
}
\description{
Drawing rainfall runoff map based on ggplot2
}
