% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Value not matching}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}
}
\value{
TRUE or FLASE
}
\description{
Value not matching
}
\examples{
1 \%!in\% 2:3
}
