% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{inverse}
\alias{inverse}
\title{Derive the inverse function}
\usage{
inverse(f, lower = 1e-08, upper = 1e+07)
}
\arguments{
\item{f}{raw function}

\item{lower}{the lower end points of the interval to be searched}

\item{upper}{the upper end points of the interval to be searched}
}
\value{
result of inverse function
}
\description{
Derive the inverse function
}
\examples{
inv = inverse(function(x) x ^ 2)
inv(4)
}
