% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_functions.R
\name{PET_Penman1948}
\alias{PET_Penman1948}
\title{Calculate potential evapotranspiration by Penman 1948 model}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Rn}{surface net radiation [W m-2]}

\item{U2}{2 meters wind speed [m s-1]}

\item{VPD}{vapor pressure deficit [kPa]}

\item{Pa}{air pressure [kPa]}
}
\value{
potential evapotranspiration calculated by Penman 1948 model [mm d-1]
}
\description{
Calculate potential evapotranspiration by Penman 1948 model
}
\examples{
PET_Penman1948(20, 50, 2, 0.5, 101)
}
