% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_functions.R
\name{PET_Penman1948_max}
\alias{PET_Penman1948_max}
\title{Calculate maximum value of potential evapotranspiration in dry environment}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Tdry}{dry environment temperature [degC]}

\item{Rn}{surface net radiation [W m-2]}

\item{U2}{2 meters wind speed [m s-1]}

\item{Pa}{air pressure [kPa]}
}
\value{
max value of potential evapotranspiration in dry environment [mm d-1]
}
\description{
Calculate maximum value of potential evapotranspiration in dry environment
}
\examples{
PET_Penman1948_max(25, 35, 50, 2, 101)
}
