% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET_PT1972.R
\name{PET_PT1972}
\alias{PET_PT1972}
\title{Calculate evapotranspiration by Priestley-Taylor (PT) model}
\usage{
PET_PT1972(Ta, Rn, Pa = 101.325, alpha = 1.26)
}
\arguments{
\item{Ta}{air temperature [degC]}

\item{Rn}{surface net radiation [W m-2]}

\item{Pa}{air pressure [kPa]}

\item{alpha}{Priestley-Taylor coefficient}

\item{Tw}{wet environment temperature [degC]}

\item{G}{soil heat flux [W m-2]}
}
\value{
evapotranspiration calculated by Priestley-Taylor model [mm d-1]
}
\description{
Calculate evapotranspiration by Priestley-Taylor (PT) model
}
\examples{
PET_PT1972(10, 50)
}
