% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET_Yang.R
\name{PET_Yang}
\alias{PET_Yang}
\title{Potential evapotranspiration (PET) calculated using
Maximum evaporating model developed by Yang and Roderick (2019)}
\usage{
PET_Yang(Rs, Rns, Rs_toa, Ta, Ts, lat, Pa = 101.325, e = 0.96)
}
\arguments{
\item{Rs}{shortwave radiation (solar radiation) [W m-2]}

\item{Rns}{net shortwave radiation [W m-2]}

\item{Rs_toa}{shortwave radiation form top of atmosphere [W m-2]}

\item{Ta}{near surface temperature (2m temperature) [degC]}

\item{Ts}{surfacee temperature [degC]}

\item{lat}{latitude [deg]}

\item{Pa}{surface pressure [kPa]}

\item{e}{surfacec emissivity}
}
\value{
potential evapotranspiration
}
\description{
Potential evapotranspiration (PET) calculated using
Maximum evaporating model developed by Yang and Roderick (2019)
}
