% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_functions.R
\name{PET_Zhou2024}
\alias{PET_Zhou2024}
\title{Potential evapotranspiration (PET) calculated using
algorithm developed by Zhou and Yu (2024)}
\arguments{
\item{Rn}{net surface radiation [W m-2]}

\item{Ta}{near surface (2m) temperature [degC]}

\item{Ts}{temperature at the evaporating surface [degC]}

\item{ea}{actual vapor pressure [kPa]}

\item{Pa}{surface pressure [kPa]}
}
\value{
potential evapotranspiration
}
\description{
Potential evapotranspiration (PET) calculated using
algorithm developed by Zhou and Yu (2024)
}
