% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_value_interp.R
\name{previous_value_interp}
\alias{previous_value_interp}
\title{Interpolate with the previous value}
\usage{
previous_value_interp(vect_data)
}
\arguments{
\item{vect_data}{raw data}
}
\value{
interpolated data
}
\description{
Interpolate with the previous value
}
\examples{
previous_value_interp(c(1, NA, 2, NA, NA, 3))
}
