% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBItools.R
\name{tidy_NCBI_info}
\alias{tidy_NCBI_info}
\title{tidy the fieldvals table of NCBI_info}
\usage{
tidy_NCBI_info(NCBI_info, sample_ID)
}
\arguments{
\item{NCBI_info}{NCBI_info}

\item{sample_ID}{sample ID}
}
\value{
a tidy data.frame
}
\description{
tidy the fieldvals table of NCBI_info
}
