% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{VPD2Td}
\alias{VPD2Td}
\title{Title Convert vapor pressure deficit to dew-point temperature}
\usage{
VPD2Td(VPD, Ta)
}
\arguments{
\item{VPD}{vapor pressure deficit [kPa]}

\item{Ta}{air temperature [degC]}
}
\value{
dew-point temperature [degC]
}
\description{
Title Convert vapor pressure deficit to dew-point temperature
}
\examples{
VPD2Td(0.5, 20)
}
