% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R, R/rcpp_functions.R
\name{cal_delta}
\alias{cal_delta}
\title{cal delta value used in ecm}
\usage{
cal_delta(x, method = "lu2025")
}
\arguments{
\item{x}{a vector}

\item{method}{method}

\item{Ta}{air temperature [degC]}
}
\value{
slope (delta) [kPa degC-1]
}
\description{
cal delta value used in ecm

Calculate the slope of the saturation vapor pressure curve at certain air
}
\examples{
cal_delta(Ta = 20)
}
