% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{cal_delta}
\alias{cal_delta}
\title{Calculate the slope of the saturation vapor pressure curve at certain air}
\usage{
cal_delta(Ta)
}
\arguments{
\item{Ta}{air temperature [degC]}
}
\value{
slope (delta) [kPa degC-1]
}
\description{
Calculate the slope of the saturation vapor pressure curve at certain air
}
\examples{
cal_delta(20)
}
