% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_ea}
\alias{cal_ea}
\title{Calculate actual vapor pressure}
\usage{
cal_ea(Ta, RH = NULL, Tdew = NULL, Pa = NULL, q = NULL)
}
\arguments{
\item{Ta}{air temperature [degC]}

\item{RH}{relative humidity [\%]}

\item{Tdew}{dew-point temperature [degC]}

\item{Pa}{pressure [kPa]}

\item{q}{specif humidity [kg kg-1]}
}
\value{
actual vapor pressure [kPa]
}
\description{
Calculate actual vapor pressure
}
\examples{
cal_ea(q = 0.5)
}
