% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{cal_es}
\alias{cal_es}
\title{Title Calculate saturation vapor pressure}
\usage{
cal_es(Ta)
}
\arguments{
\item{Ta}{air temperature [degC]}
}
\value{
saturation vapor pressure [kPa]
}
\description{
Title Calculate saturation vapor pressure
}
\examples{
cal_es(20)
}
