% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_functions.R
\name{cal_es}
\alias{cal_es}
\title{Calculate saturation vapor pressure}
\arguments{
\item{Ta}{air temperature [degC]}
}
\value{
saturation vapor pressure [kPa]
}
\description{
Calculate saturation vapor pressure
}
\examples{
cal_es(20)
}
