% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_gma}
\alias{cal_gma}
\title{Calculate psychrometric constant}
\usage{
cal_gma(Pa = NULL, lambda = NULL)
}
\arguments{
\item{Pa}{pressure [kPa]}

\item{lambda}{lantent heat of vaporization [MJ kg-1]}
}
\value{
psychrometric constant [kPa degC-1]
}
\description{
Calculate psychrometric constant
}
\examples{
cal_gma()
}
