% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_functions.R
\name{cal_gma}
\alias{cal_gma}
\title{Calculate psychrometric constant}
\arguments{
\item{Pa}{air pressure [kPa]}

\item{Ta}{lantent heat of vaporization [MJ kg-1]}
}
\value{
psychrometric constant [kPa degC-1]
}
\description{
Calculate psychrometric constant
}
\examples{
cal_gma(Pa = 100, Ta = 20)
}
