% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{cal_lambda}
\alias{cal_lambda}
\title{Title Calculate lantent heat of vaporization}
\usage{
cal_lambda(Ta = NULL)
}
\arguments{
\item{Ta}{air temperature [degC]}
}
\value{
lambda, latent heat of vaporization, about [2.45 MJ kg-1]
}
\description{
Title Calculate lantent heat of vaporization
}
\examples{
cal_lambda(20)
}
