% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{cal_Tdry}
\alias{cal_Tdry}
\title{Calculate dry environment temperature}
\usage{
cal_Tdry(Twb, Pa = 101.325, Ta = NULL)
}
\arguments{
\item{Twb}{wet-bulb temperature [degC]}

\item{Pa}{air pressure [kPa]}

\item{Ta}{air temperature [degC]}
}
\value{
dry environment temperature [degC]
}
\description{
Calculate dry environment temperature
}
\examples{
cal_Tdry(5)
}
