% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_Twb}
\alias{cal_Twb}
\title{Calculate wet-bulb temperature}
\usage{
cal_Twb(Td, Ta, Pa = NULL, gma = NULL)
}
\arguments{
\item{Td}{dew-point temperature [degC]}

\item{Ta}{air temperature [degC]}

\item{Pa}{pressure [kPa]}

\item{gma}{psychrometric constant [kPa degC-1]}
}
\value{
wet-bulb temperature [degC]
}
\description{
Calculate wet-bulb temperature
}
\examples{
cal_Twb(10, 20)
}
