% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ET_by_CR_model.R
\name{cal_Tws}
\alias{cal_Tws}
\title{Calculate wet surface temperature}
\usage{
cal_Tws(Rn, Ep, Ta, ea, Pa = NULL, gma = NULL, G = NULL)
}
\arguments{
\item{Rn}{net radiation [W m-2]}

\item{Ep}{potential evapotranspiration [mm d-1]}

\item{Ta}{air temperature [degC]}

\item{ea}{actual vapor pressure [kPa]}

\item{Pa}{pressure [kPa]}

\item{gma}{psychrometric constant [kPa degC-1]}

\item{G}{soil heat flux [W m-2]}
}
\value{
wet surface temperature [degC]
}
\description{
Calculate wet surface temperature
}
\examples{
cal_Tws(Rn = 50, Ep = 5, Ta = 20, ea = 0.3)
}
