% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_params.R
\name{cal_U2}
\alias{cal_U2}
\title{Title Converted zm wind speed to 2m wind speed}
\usage{
cal_U2(Uz, z.wind = 10)
}
\arguments{
\item{Uz}{z meters wind speed [m s-1]}

\item{z.wind}{measurement altitude of wind speed [m]}
}
\value{
2 meters wind speed [m s-1]
}
\description{
Title Converted zm wind speed to 2m wind speed
}
\examples{
cal_U2(6)
}
