% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBItools.R
\name{get_NCBI_info}
\alias{get_NCBI_info}
\title{get NCBI sample information table}
\usage{
get_NCBI_info(sample_ID, tidy = T)
}
\arguments{
\item{sample_ID}{sample ID}

\item{tidy}{whether tidy the fieldvals table}
}
\value{
NCBI_info (list or data.frame)
}
\description{
get NCBI sample information table
}
\examples{
get_NCBI_info(sample_ID = 'ERX2746395')
}
