% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRunoff.R
\name{ggRunoff}
\alias{ggRunoff}
\title{Drawing rainfall runoff map based on ggplot2}
\usage{
ggRunoff(d, date, Q, prcp, ymax = NULL, prcp_color = NULL)
}
\arguments{
\item{d}{input data [data.frame]}

\item{date}{Column name of date [character]}

\item{Q}{column name of runoff [character]}

\item{prcp}{column name of precipitation [character]}

\item{ymax}{Y-axis maximum, also second abscissa axis position}

\item{prcp_color}{fill color for rainfall}
}
\value{
rainfall runoff map based on ggplot2
}
\description{
Drawing rainfall runoff map based on ggplot2
}
\examples{
 d <- data.frame(date = seq(as.POSIXct('2022-05-02'),
                        as.POSIXct('2022-05-05'), by = 'day'),
                 Q    = seq(10, 40, 10),
                 P    = c(1, 2, 0, 2))
 ggRunoff(d, date = 'date', Q = 'Q', prcp = 'P')
}
