% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET_Penman1948.R
\name{PET_Penman1948_max}
\alias{PET_Penman1948_max}
\title{Calculate maximum value of potential evapotranspiration in dry environment}
\usage{
PET_Penman1948_max(Tdry, Rn, U2, Pa = 101.325, Ta = NULL, G = NULL)
}
\arguments{
\item{Tdry}{dry environment temperature [degC]}

\item{Rn}{surface net radiation [W m-2]}

\item{U2}{2 meters wind speed [m s-1]}

\item{Pa}{air pressure [kPa]}

\item{Ta}{air temperature [degC]}

\item{G}{soil heat flux [W m-2]}
}
\value{
max value of potential evapotranspiration in dry environment [mm d-1]
}
\description{
Calculate maximum value of potential evapotranspiration in dry environment
}
\examples{
PET_Penman1948_max(35, 50, 2)
}
