% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_functions.R
\name{PET_Yang2019}
\alias{PET_Yang2019}
\title{Potential evapotranspiration (PET) calculated using
Maximum evaporating model developed by Yang and Roderick (2019)}
\arguments{
\item{Rs}{shortwave radiation (solar radiation) [W m-2]}

\item{Rns}{net shortwave radiation [W m-2]}

\item{Rs_toa}{shortwave radiation form top of atmosphere [W m-2]}

\item{Ts}{temperature at the evaporating surface [degC]}

\item{lat}{latitude [deg]}

\item{Pa}{surface pressure [kPa]}
}
\value{
potential evapotranspiration
}
\description{
Potential evapotranspiration (PET) calculated using
Maximum evaporating model developed by Yang and Roderick (2019)
}
